Move authenticator to util package
