+++ pom.xml
@@ -194,6 +194,7 @@
                                 <exclude>com/quorum/tessera/data/migration/Main*</exclude>
                                 <exclude>com/quorum/tessera/config/migration/Main*</exclude>
                                 <exclude>com/quorum/tessera/config/util/PasswordReader*</exclude>
+                                <exclude>com/quorum/tessera/util/KeyVaultAuthenticator*</exclude>
                             </excludes>
                             
                             <rules>
