+++ tessera-app/src/main/resources/tessera-spring.xml
@@ -137,7 +137,7 @@
 
     <bean id="keyConfig" factory-bean="config" factory-method="getKeys" />
 
-    <bean id="keyVaultClient" class="com.quorum.tessera.key.vault.KeyVaultAuthenticator" factory-method="getAuthenticatedClient"/>
+    <bean id="keyVaultClient" class="com.quorum.tessera.util.KeyVaultAuthenticator" factory-method="getAuthenticatedClient"/>
 
     <bean id="keyVaultService" class="com.quorum.tessera.key.vault.KeyVaultService">
         <constructor-arg ref="keyConfig"/>
